+++ dc/azext_dc/_params.py
@@ -3,17 +3,46 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
-# from knack.arguments import CLIArgumentType
+from knack.arguments import CLIArgumentType
 
 from ._validators import dc_source_version_validator, user_validator
 
 
 def load_arguments(self, _):
+
+    confirm_type = CLIArgumentType(
+        help='Do not prompt for confirmation. WARNING: This is irreversible.',
+        action='store_true',
+        required=False,
+        options_list=['--yes', '-y']
+    )
+
     with self.argument_context('dc upgrade') as c:
         c.argument('version', options_list=['--version', '-v'], help='Version (tag). Default: latest stable.',
                    validator=dc_source_version_validator)
         c.argument('prerelease', options_list=['--pre'], action='store_true',
-                   help='Deploy latest prerelease version.')
+                   help="The id of the user. If value is 'me', the identity is taken from the authentication "
+                   "context.")
+
+    with self.argument_context('dc user check') as c:
+        c.argument('user', options_list=['--user', '-u'],
+                   help='User ')
+
+    with self.argument_context('dc box') as c:
+        c.argument('project_names', options_list=['--projects'], nargs='*',
+                   help='Space-separated Project names to include. If not specified, all projects are included.')
+
+    for scope in ['dc box', 'dc project box']:
+        with self.argument_context(scope) as c:
+            c.argument('dev_center', options_list=['--dev-center', '-dc'], help='Dev Center name.')
+            c.argument('pool_names', options_list=['--pools'], nargs='*',
+                       help='Space-separated Pool names to include. If not specified, all pools are included.')
+            c.argument('user_ids', options_list=['--users'], nargs='*',
+                       help='Space-separated user ids to include. If not specified, all users are included.')
+
+    for scope in ['dc box stop', 'dc box start', 'dc project box stop', 'dc project box start']:
+        with self.argument_context(scope) as c:
+            c.argument('yes', arg_type=confirm_type)
 
     with self.argument_context('dc project') as c:
         # c.argument('dev_center', options_list=['--dev-center-name', '--dev-center', '-dc'], help='Project prefix.')
@@ -24,3 +53,7 @@ def load_arguments(self, _):
             c.argument('user_id', options_list=['--user-id', '--user', '-u'], default='me', validator=user_validator,
                        help="The id of the user. If value is 'me', the identity is taken from the authentication "
                        "context.")
+
+    with self.argument_context('dc project user add') as c:
+        c.argument('skip_license_check', options_list=['--skip-license-check'], action='store_true',
+                   help='Skip license check.')