+++ dc/azext_dc/commands.py
@@ -13,6 +13,20 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
 
     with self.command_group('dc') as g:
         g.custom_command('upgrade', 'dc_upgrade')
+        # g.custom_command('test', 'dc_test')
+
+    with self.command_group('dc user') as g:
+        g.custom_command('check', 'dc_user_check')
+
+    with self.command_group('dc box') as g:
+        g.custom_command('list', 'list_dc_boxes')
+        g.custom_command('start', 'start_dc_boxes')
+        g.custom_command('stop', 'stop_dc_boxes')
+
+    with self.command_group('dc project box') as g:
+        g.custom_command('list', 'list_project_boxes')
+        g.custom_command('start', 'start_project_boxes')
+        g.custom_command('stop', 'stop_project_boxes')
 
     with self.command_group('dc project user') as g:
         g.custom_command('add', 'add_project_user')