+++ dc/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.0.2
+++++++
++ list, start, stop boxes at devcenter and project level
+* check if user has the licenses required for dev box
+
 0.0.1
 ++++++
 + Initial Release